requirements.txt
google-adk==0.5.0
google-genai==1.15.0
python-dotenv==1.1.0
protobuf~=6.31.0